# openapi.yaml
openapi: 3.0.1
info:
  title: Acuarium Management API
  version: 2.0.0
  description: |
    API REST para la gestión de un acuario: peces, plantas, iluminación y mantenimiento.

    Desarrollada con **SDD (Spec Driven Development)**: este archivo es la fuente de verdad
    del proyecto. Todo el código — backend y frontend — se construye a partir de este contrato.

    ## Autenticación
    La API usa **JWT Bearer tokens**. Para acceder a los endpoints protegidos:
    1. Obtén un token haciendo `POST /auth/login`
    2. Incluye el token en la cabecera: `Authorization: Bearer <token>`

    ## Roles
    - `ROLE_ADMIN` → acceso total (CRUD completo + gestión de usuarios + DELETE en todas las entidades)
    - `ROLE_USER`  → lectura y escritura sobre sus propios recursos (GET / POST / PUT)

    ### Tabla de permisos por endpoint

    | Método | Endpoint          | ROLE_USER  | ROLE_ADMIN |
    |--------|-------------------|------------|------------|
    | POST   | /auth/register    | ✗          | ✓          |
    | POST   | /auth/login       | público    | público    |
    | POST   | /auth/refresh     | público    | público    |
    | GET    | /auth/me          | ✓          | ✓          |
    | GET    | /{recurso}        | ✓          | ✓          |
    | POST   | /{recurso}        | ✓          | ✓          |
    | GET    | /{recurso}/{id}   | ✓          | ✓          |
    | PUT    | /{recurso}/{id}   | ✓ (propio) | ✓          |
...

    ## Versiones
...
servers:
...

...
security:
  - bearerAuth: []
...
paths:
...
  /plants:
    get:
      tags: [plants]
      summary: Listar todas las plantas
      operationId: listPlants
      description: |
        **Roles requeridos:** `ROLE_USER` | `ROLE_ADMIN`
      x-required-roles: [ROLE_USER, ROLE_ADMIN]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - in: query
          name: name
          schema:
            type: string
          description: Filtrar por nombre (búsqueda parcial)
        - in: query
          name: species
          schema:
            type: string
          description: Filtrar por especie
        - in: query
          name: acuarioId
          schema:
            type: integer
          description: Filtrar por acuario
        - in: query
          name: sort
          schema:
            type: string
            enum: [name, species, age]
            default: name
          description: Campo por el que ordenar
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Dirección del ordenamiento
      responses:
        '200':
          description: Lista de plantas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plant'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [plants]
      summary: Añadir una nueva planta
      operationId: createPlant
      description: |
        **Roles requeridos:** `ROLE_USER` | `ROLE_ADMIN`
      x-required-roles: [ROLE_USER, ROLE_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantCreate'
            example:
              name: "Anubias"
              species: "Anubias barteri"
              age: 1
              acuarioId: 1
      responses:
        '201':
          description: Planta creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /plants/{id}:
    parameters:
      - $ref: '#/components/parameters/idParam'
    get:
      tags: [plants]
      summary: Obtener una planta por ID
      operationId: getPlant
      description: |
        **Roles requeridos:** `ROLE_USER` | `ROLE_ADMIN`
      x-required-roles: [ROLE_USER, ROLE_ADMIN]
      responses:
        '200':
          description: Detalles de la planta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [plants]
      summary: Actualizar una planta
      operationId: updatePlant
      description: |
        `ROLE_USER` solo puede actualizar sus propias plantas.

        **Roles requeridos:** `ROLE_USER` (propio) | `ROLE_ADMIN`
      x-required-roles: [ROLE_USER, ROLE_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantCreate'
      responses:
        '200':
          description: Planta actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [plants]
      summary: Eliminar una planta
      operationId: deletePlant
      description: |
        **Roles requeridos:** `ROLE_ADMIN`
      x-required-roles: [ROLE_ADMIN]
      responses:
        '204':
          description: Planta eliminada correctamente
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ----------------------------------------------------------
  # LIGHTING
  # ----------------------------------------------------------
  /lighting:
    get:
      tags: [lighting]
      summary: Listar todas las iluminaciones
      operationId: listLighting
      description: |
        **Roles requeridos:** `ROLE_USER` | `ROLE_ADMIN`
      x-required-roles: [ROLE_USER, ROLE_ADMIN]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - in: query
          name: light_type
          schema:
            type: string
          description: Filtrar por tipo de luz (LED, fluorescente...)
        - in: query
          name: wifi
          schema:
            type: boolean
          description: Filtrar por conectividad wifi
        - in: query
          name: aquariumId
          schema:
            type: integer
          description: Filtrar por acuario
        - in: query
          name: sort
          schema:
            type: string
            enum: [light_type, watts, start_time]
            default: light_type
          description: Campo por el que ordenar
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Dirección del ordenamiento
      responses:
        '200':
          description: Lista de iluminaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lighting'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [lighting]
      summary: Añadir una nueva iluminación
      operationId: createLighting
      description: |
        **Roles requeridos:** `ROLE_USER` | `ROLE_ADMIN`
      x-required-roles: [ROLE_USER, ROLE_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LightingCreate'
            example:
              light_type: "LED"
              watts: 20
              wifi: true
              start_time: "08:00"
              end_time: "20:00"
              aquariumId: 1
      responses:
        '201':
          description: Iluminación creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lighting'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /lighting/{id}:
    parameters:
      - $ref: '#/components/parameters/idParam'
    get:
      tags: [lighting]
      summary: Obtener una iluminación por ID
      operationId: getLighting
      description: |
        **Roles requeridos:** `ROLE_USER` | `ROLE_ADMIN`
      x-required-roles: [ROLE_USER, ROLE_ADMIN]
      responses:
        '200':
          description: Detalles de la iluminación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lighting'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [lighting]
      summary: Actualizar una iluminación
      operationId: updateLighting
      description: |
        `ROLE_USER` solo puede actualizar sus propias iluminaciones.

        **Roles requeridos:** `ROLE_USER` (propio) | `ROLE_ADMIN`
      x-required-roles: [ROLE_USER, ROLE_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LightingCreate'
      responses:
        '200':
          description: Iluminación actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lighting'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [lighting]
      summary: Eliminar una iluminación
      operationId: deleteLighting
      description: |
        **Roles requeridos:** `ROLE_ADMIN`
      x-required-roles: [ROLE_ADMIN]
      responses:
        '204':
          description: Iluminación eliminada correctamente
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ----------------------------------------------------------
  # MAINTENANCE
  # ----------------------------------------------------------
  /maintenance:
    get:
      tags: [maintenance]
      summary: Listar tareas de mantenimiento
      operationId: listMaintenance
      description: |
        **Roles requeridos:** `ROLE_USER` | `ROLE_ADMIN`
      x-required-roles: [ROLE_USER, ROLE_ADMIN]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - in: query
          name: task
          schema:
            type: string
          description: Filtrar por nombre de tarea (búsqueda parcial)
        - in: query
          name: date
          schema:
            type: string
            format: date
          description: Filtrar por fecha (YYYY-MM-DD)
        - in: query
          name: acuarioId
          schema:
            type: integer
          description: Filtrar por acuario
        - in: query
          name: sort
          schema:
            type: string
            enum: [task, date]
            default: date
          description: Campo por el que ordenar
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Dirección del ordenamiento
      responses:
        '200':
          description: Lista de tareas de mantenimiento
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Maintenance'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [maintenance]
      summary: Crear una nueva tarea de mantenimiento
      operationId: createMaintenance
      description: |
        **Roles requeridos:** `ROLE_USER` | `ROLE_ADMIN`
      x-required-roles: [ROLE_USER, ROLE_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceCreate'
            example:
              task: "Cambio de agua parcial 30%"
              date: "2026-03-16T10:00:00"
              acuarioId: 1
      responses:
        '201':
          description: Tarea de mantenimiento creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Maintenance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /maintenance/{id}:
    parameters:
      - $ref: '#/components/parameters/idParam'
    get:
      tags: [maintenance]
      summary: Obtener una tarea de mantenimiento por ID
      operationId: getMaintenance
      description: |
        **Roles requeridos:** `ROLE_USER` | `ROLE_ADMIN`
      x-required-roles: [ROLE_USER, ROLE_ADMIN]
      responses:
        '200':
          description: Detalles de la tarea de mantenimiento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Maintenance'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [maintenance]
      summary: Actualizar una tarea de mantenimiento
      operationId: updateMaintenance
      description: |
        `ROLE_USER` solo puede actualizar sus propias tareas.

        **Roles requeridos:** `ROLE_USER` (propio) | `ROLE_ADMIN`
      x-required-roles: [ROLE_USER, ROLE_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceCreate'
      responses:
        '200':
          description: Tarea de mantenimiento actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Maintenance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [maintenance]
      summary: Eliminar una tarea de mantenimiento
      operationId: deleteMaintenance
      description: |
        **Roles requeridos:** `ROLE_ADMIN`
      x-required-roles: [ROLE_ADMIN]
      responses:
        '204':
          description: Tarea de mantenimiento eliminada correctamente
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'


# ============================================================
# COMPONENTS
# ============================================================
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido en `POST /auth/login`.
        Incluirlo en la cabecera: `Authorization: Bearer <token>`

        - **access_token** expira en 30 minutos
        - **refresh_token** expira en 7 días (usar `POST /auth/refresh`)

  # ----------------------------------------------------------
  # Reusable parameters
  # ----------------------------------------------------------
  parameters:
    idParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
        minimum: 1
      description: Identificador único del recurso
    limitParam:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Número máximo de resultados a devolver
    offsetParam:
      in: query
      name: offset
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Número de resultados a omitir (paginación)

  # ----------------------------------------------------------
  # Reusable responses
  # ----------------------------------------------------------
  responses:
    Unauthorized:
      description: Token ausente, inválido o expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Not authenticated"
    Forbidden:
      description: Sin permisos suficientes para esta operación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Not enough permissions"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Resource not found"
    BadRequest:
      description: Datos de entrada inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Invalid input data"
    UnprocessableEntity:
      description: Error de validación en los campos del body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'

  # ----------------------------------------------------------
  # Schemas
  # ----------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      description: Respuesta de error genérica
      properties:
        detail:
          type: string
          description: Descripción del error
          example: "Resource not found"
      required:
        - detail
    ValidationErrorResponse:
      type: object
      description: Errores de validación de Pydantic/FastAPI
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
                example: ["body", "email"]
              msg:
                type: string
                example: "value is not a valid email address"
              type:
                type: string
                example: "value_error.email"
      required:
        - detail

    # ---- Auth ----
...
    Plant:
      type: object
      description: Planta con todos sus datos
      properties:
        id:
          type: integer
          readOnly: true
          example: 3
        name:
          type: string
          minLength: 2
          maxLength: 50
          example: "Anubias"
        species:
          type: string
          minLength: 2
          maxLength: 50
          example: "Anubias barteri"
        age:
          type: integer
          minimum: 0
          maximum: 10
          example: 1
        picture:
          type: string
          format: uri
          nullable: true
          example: "https://example.com/anubias.jpg"
        description:
          type: string
          maxLength: 200
          nullable: true
          example: "Planta resistente y de bajo mantenimiento."
        acuarioId:
          type: integer
          description: ID del acuario al que pertenece la planta
          example: 1
      required:
        - name
        - species
        - age
        - acuarioId
    PlantCreate:
      type: object
      description: Datos para crear o actualizar una planta
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
          example: "Anubias"
        species:
          type: string
          minLength: 2
          maxLength: 50
          example: "Anubias barteri"
        age:
          type: integer
          minimum: 0
          maximum: 10
          example: 1
        picture:
          type: string
          format: uri
          nullable: true
          example: "https://example.com/anubias.jpg"
        description:
          type: string
          maxLength: 200
          nullable: true
          example: "Planta resistente y de bajo mantenimiento."
        acuarioId:
          type: integer
          example: 1
      required:
        - name
        - species
        - age
        - acuarioId
    Lighting:
      type: object
      description: Iluminacion con todos sus datos
      properties:
        id:
          type: integer
          readOnly: true
          example: 2
        light_type:
          type: string
          minLength: 2
          maxLength: 30
          description: Tipo de luz (LED, fluorescente, haluro metalico...)
          example: "LED"
        watts:
          type: integer
          minimum: 1
          maximum: 1000
          description: Potencia en vatios
          example: 20
        wifi:
          type: boolean
          description: Dispone de conectividad wifi
          example: true
        start_time:
          type: string
          pattern: '^([01]\d|2[0-3]):[0-5]\d$'
          description: Hora de encendido (HH:mm)
          example: "08:00"
        end_time:
          type: string
          pattern: '^([01]\d|2[0-3]):[0-5]\d$'
          description: Hora de apagado (HH:mm)
          example: "20:00"
        aquariumId:
          type: integer
          description: ID del acuario al que pertenece
          example: 1
      required:
        - light_type
        - watts
        - wifi
        - start_time
        - end_time
        - aquariumId
    LightingCreate:
      type: object
      description: Datos para crear o actualizar una iluminacion
      properties:
        light_type:
          type: string
          minLength: 2
          maxLength: 30
          example: "LED"
        watts:
          type: integer
          minimum: 1
          maximum: 1000
          example: 20
        wifi:
          type: boolean
          example: true
        start_time:
          type: string
          pattern: '^([01]\d|2[0-3]):[0-5]\d$'
          example: "08:00"
        end_time:
          type: string
          pattern: '^([01]\d|2[0-3]):[0-5]\d$'
          example: "20:00"
        aquariumId:
          type: integer
          example: 1
      required:
        - light_type
        - watts
        - wifi
        - start_time
        - end_time
        - aquariumId
    Maintenance:
      type: object
      description: Tarea de mantenimiento con todos sus datos
      properties:
        id:
          type: integer
          readOnly: true
          example: 10
        task:
          type: string
          minLength: 3
          maxLength: 100
          example: "Cambio de agua parcial 30%"
        date:
          type: string
          format: date-time
          example: "2026-03-16T10:00:00"
        acuarioId:
          type: integer
          description: ID del acuario al que pertenece la tarea
          example: 1
      required:
        - task
        - date
        - acuarioId
    MaintenanceCreate:
      type: object
      description: Datos para crear o actualizar una tarea de mantenimiento
      properties:
        task:
          type: string
          minLength: 3
          maxLength: 100
          example: "Cambio de agua parcial 30%"
        date:
          type: string
          format: date-time
          example: "2026-03-16T10:00:00"
        acuarioId:
          type: integer
          example: 1
      required:
        - task
        - date
        - acuarioId